fix(api): delete dead subscription routes/service, extract working API-key routes
The Wave 115 subscription billing surface was dead code: subscriptionService.ts
was @ts-nocheck, instantiated its own PrismaClient, and queried prisma.subscription.*
— but there is no `model Subscription` in the schema (only SubscriptionApiKey).
Every path through POST/GET/DELETE /api/subscriptions* 500'd for all valid inputs.
There is no client (the only billing UI is archived under _archive/wave119) and
subscriptions are not a live product surface.

- Delete services/billing/subscriptionService.ts and routes/subscriptions.ts.
- Extract the working API-key handlers (backed by the real SubscriptionApiKey
  model, used by rateLimiter) into routes/apiKeys.ts, preserving PR #501's
  uuid-format guard on DELETE /api/api-keys/:keyId.
- app.ts: register registerApiKeyRoutes in place of registerSubscriptionRoutes.
- contracts.ts: drop orphaned createSubscriptionRequestSchema / CreateSubscriptionRequest
  (subscriptionTierSchema stays — still used by the API-key schema).
- docs.ts: remove the /api/subscriptions OpenAPI path + summary entry so published
  docs stop advertising a 500-ing endpoint.

Design Handoff References:
> No Claude Design handoff file used for this PR.

Verification:
- prisma generate: green
- backend tsc --noEmit: green (proves nothing else imports the deleted modules)
- repo-wide sweep for subscriptionService / registerSubscriptionRoutes /
  createSubscriptionRequestSchema / /api/subscriptions: no remaining references
- OpenAPI docs no longer advertise /api/subscriptions
- API-key routes still registered (registerApiKeyRoutes in app.ts)
