+++ apps/api/backend/src/routes/apiKeys.ts
@@ -1,0 +1,68 @@
+/**
+ * apiKeys.ts — Wave 115: API Key Routes
+ *
+ * POST   /api/api-keys                         — Generate API key
+ * GET    /api/api-keys/:clinicianId             — List keys for clinician
+ * DELETE /api/api-keys/:keyId                  — Revoke key
+ */
+
+import type { Express, Request, Response } from 'express';
+import { createApiKeyRequestSchema } from '../services/billing/contracts';
+import { generateApiKey, revokeApiKey, getApiKeysByClinicianId } from '../services/billing/apiKeyService';
+import { log } from '../obs/logger';
+
+function readValidationMessage(error: { issues?: Array<{ message?: string }> }): string {
+  return error.issues?.[0]?.message ?? 'Invalid request payload.';
+}
+
+// SubscriptionApiKey.id is a Postgres uuid column — querying it with a
+// non-uuid string makes Prisma throw (a 500) instead of returning null.
+const UUID_RE = /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i;
+
+export function registerApiKeyRoutes(app: Express): void {
+
+  // ── POST /api/api-keys ──────────────────────────────────────────
+  app.post('/api/api-keys', async (req: Request, res: Response) => {
+    try {
+      const parsed = createApiKeyRequestSchema.safeParse(req.body ?? {});
+      if (!parsed.success) {
+        res.status(400).json({ error: readValidationMessage(parsed.error) });
+        return;
+      }
+
+      const { clinicianId, name, tier } = parsed.data;
+      const result = await generateApiKey(clinicianId, name, tier);
+      res.status(201).json(result);
+    } catch (err) {
+      const msg = err instanceof Error ? err.message : 'Unknown error';
+      log('error', 'api_key_create_error', { error: msg });
+      res.status(500).json({ error: msg });
+    }
+  });
+
+  // ── GET /api/api-keys/:clinicianId ──────────────────────────────
+  app.get('/api/api-keys/:clinicianId', async (req: Request, res: Response) => {
+    try {
+      const keys = await getApiKeysByClinicianId(req.params.clinicianId);
+      res.json(keys);
+    } catch (err) {
+      const msg = err instanceof Error ? err.message : 'Unknown error';
+      res.status(500).json({ error: msg });
+    }
+  });
+
+  // ── DELETE /api/api-keys/:keyId ─────────────────────────────────
+  app.delete('/api/api-keys/:keyId', async (req: Request, res: Response) => {
+    try {
+      if (!UUID_RE.test(req.params.keyId ?? '')) {
+        res.status(404).json({ error: 'API key not found' });
+        return;
+      }
+      await revokeApiKey(req.params.keyId);
+      res.json({ revoked: true });
+    } catch (err) {
+      const msg = err instanceof Error ? err.message : 'Unknown error';
+      res.status(500).json({ error: msg });
+    }
+  });
+}
