+++ apps/api/backend/src/services/billing/contracts.ts
@@ -8,19 +8,12 @@ import {
 
 export const subscriptionTierSchema = z.enum(BILLING_TIERS);
 
-export const createSubscriptionRequestSchema = z.object({
-  clinicianId: z.string().trim().min(1).max(120),
-  tier: subscriptionTierSchema,
-  stripeSubId: z.string().trim().min(1).max(255).optional(),
-});
-
 export const createApiKeyRequestSchema = z.object({
   clinicianId: z.string().trim().min(1).max(120),
   name: z.string().trim().min(1).max(120),
   tier: subscriptionTierSchema.optional().default('STARTER'),
 });
 
-export type CreateSubscriptionRequest = z.infer<typeof createSubscriptionRequestSchema>;
 export type CreateApiKeyRequest = z.infer<typeof createApiKeyRequestSchema>;
 
 export type { BillingFeature, BillingPlan, BillingTier };
